refactor: split bg requests.js
